fix some loader issues
