+++ packages/engine/tests/util/loadGLTFAssetNode.ts
@@ -4,7 +4,7 @@ import path from 'path'
 
 import { createGLTFLoader } from '../../src/assets/functions/createGLTFLoader'
 import { GLTF } from '../../src/assets/loaders/gltf/GLTFLoader'
-import './patchBrowserForNode.ts'
+import '../../src/patchEngineNode.ts'
 
 const toArrayBuffer = (buf) => {
   const arrayBuffer = new ArrayBuffer(buf.length)
